+++ Changes.md
@@ -2,13 +2,24 @@
 
 *Table of Contents*
 
-- [0.29.1 (2024-05-15)](#0291-2024-05-15)
+- [Unreleased](#unreleased)
   - [✨ Enhancements](#-enhancements)
+- [0.29.1 (2024-05-15)](#0291-2024-05-15)
+  - [✨ Enhancements](#-enhancements-1)
 - [0.29.0 (2024-05-07)](#0290-2024-05-07)
   - [🚨 Breaking changes](#-breaking-changes)
-  - [✨ Enhancements](#-enhancements-1)
+  - [✨ Enhancements](#-enhancements-2)
   - [🚧 Internal changes](#-internal-changes)
 - [Prior to 0.29](#prior-to-029)
+
+
+## Unreleased
+
+### ✨ Enhancements
+
+* Introduced `config parse` command to parses and displays the configuration file's contents without executing any commands.
+
+  This command helps users understand how secrets.env interprets the configuration file and identifies critical errors.
 
 
 ## 0.29.1 (2024-05-15)